requirements.txt
streamlit>=1.32.0
google-genai>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
kaleido>=0.2.1
reportlab>=4.0.0
pdfplumber>=0.10.0
scikit-learn>=1.3.0
openpyxl>=3.1.0
